% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newsboard_data.R
\name{newsboard_data}
\alias{newsboard_data}
\title{The Swiss Stock Exchange Market Data News via the SIX Group Newsboard}
\usage{
newsboard_data(
  firstDate = Sys.Date() - 365 * 2,
  lastDate = Sys.Date(),
  messageType = c("Emergency Information", "Mistrade", "Suspension",
    "Trading Information"),
  Market = c("Six Swiss Exchange", "Structured Products",
    "XBTR - Bilateral Trading Plattform"),
  Products = c("Bonds", "Derivatives", "Equities", "Exchange Traded Products", "Funds")
)
}
\arguments{
\item{firstDate}{a character string indicating the year-month-date of From date.
The default is equivalent to firstDate = \code{Sys.Date() - 365*2}, the earliest date for which
data is available on the Newsboard.}

\item{lastDate}{a character string indicating the year-month-date of To date.
The default is equivalent to `lastDate = Sys.Date()``, the last date in range for which
data is available, but not greater than current date on the Newsboard.}

\item{messageType}{a vector of character strings of the message type.
This must be one or combination of either of the strings "Emergency Information",
"Mistrade", "Suspension", "Trading Information".The default is
equivalent to messageType = 'All'.}

\item{Market}{a vector of character strings of the Market type.
This must be one or combination of either of the strings "Six Swiss Exchange",
"Structured Products","XBTR - Bilateral Trading Plattform". The default is
equivalent to Market = 'All'.}

\item{Products}{a vector of character strings of the products type.
This must be one or combination of either of the strings "Bonds", "Derivatives",
"Equities", "Exchange Traded Products", "Funds".The default is
equivalent to Products = 'All'.}
}
\value{
Swiss Stock Exchange Market Data News via the SIX Group Newsboard
}
\description{
Returns the Swiss Stock Exchange Market Data News via the SIX Group Newsboard
}
\details{
The inputs must all be character strings, with the date format as \%Y-\%m-\%d
(as determined by \code{as.Date()}). The 'messageType', 'Market', 'Products', if not
supplied, all character string options are used, which leads to the
messageType = 'All', Market = 'All' and Products = 'All' defaults.
Supplying individual parameter requires the use of complete string
as opposed to the use of abbreviations.
}
\examples{
library(plyr)
library(jsonlite)

data_final <- newsboard_data(firstDate = '2023-06-24')
head(data_final, 1)
data_final <- newsboard_data(firstDate = '2023-06-24', messageType = 'Mistrade')
head(data_final, 2)
data_final <- newsboard_data(firstDate = '2023-06-24', Market = 'Six Swiss Exchange')
head(data_final,1)
data_final <- newsboard_data(firstDate = '2023-06-24', Products = 'Funds')
head(data_final, 1)
\dontrun{
data_final <- newsboard_data(lastDate = '2021-07-01') # error when lastdate > firstdate
head(data_final, 1)
}




match
}
\references{
Newsboard: Market activity at the Swiss Stock Exchange. SIX. https://www.six-group.com/en/products-services/the-swiss-stock-exchange/market-data/news-tools/newsboard.html#/
}
\author{
Akhapwoya Obuchere
}
